<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
=======
% Generated by roxygen2 (4.1.0.9000): do not edit by hand
>>>>>>> a538c1a4fc0b3ebc22dbed7ead515583fd172a6e
% Please edit documentation in R/RcppExports.R
\name{boundingBoxFromCIDR}
\alias{boundingBoxFromCIDR}
\title{Bounding box from CIDR blocks}
\usage{
boundingBoxFromCIDR(cidr)
}
\arguments{
\item{cidr}{character vector of dotted-decimal/digit CIDRs}
}
\value{
list of bounding box extents for each CIDR - \code{cidr},
        \code{xmin}, \code{ymin}, \code{xmax}, \code{ymax}
}
\description{
Returns a \code{list} of bounding boxes for a given CIDR within the
Hilbert-curve plane
}
\examples{
boundingBoxFromCIDR("30.0.0.0/8")
}

