% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ip2long}
\alias{ip2long}
\title{Convert IP addresses in dotted-decimal notation to long integers}
\usage{
ip2long(ip)
}
\arguments{
\item{ip}{input character vector}
}
\value{
vector of equivalent long integer IP addresses
}
\description{
Convert IP addresses in dotted-decimal notation to long integers
}

