% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ipv4heatmap-package.R
\docType{package}
\name{ipv4heatmap-package}
\alias{ipv4heatmap-package}
\title{A package to create curve heatmaps of IPv4 addresses}
\description{
This package provides functions for taking a vector of
IPv4 addresses and building a Hilbert curve from them.
}
\seealso{
\url{http://xkcd.com/195/},
         \url{http://en.wikipedia.org/wiki/Hilbert_curve},
         \url{http://maps.measurement-factory.com/}
}

