<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
=======
% Generated by roxygen2 (4.1.0.9000): do not edit by hand
>>>>>>> a538c1a4fc0b3ebc22dbed7ead515583fd172a6e
% Please edit documentation in R/ipv4heatmap-package.R
\docType{package}
\name{ipv4heatmap-package}
\alias{ipv4heatmap-package}
\title{A package to create curve heatmaps of IPv4 addresses}
\description{
This package provides functions for taking a vector of
IPv4 addresses and building a Hilbert curve from them.
}
\seealso{
\url{http://xkcd.com/195/},
         \url{http://en.wikipedia.org/wiki/Hilbert_curve},
         \url{http://maps.measurement-factory.com/}
}

