% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ipv4heatmap-package}
\alias{ipv4heatmap-package}
\title{Creates Hilbert curve heatmaps of IPv4 addresses}
\description{
Takes a vector of character (dotted-decimal) IPv4 addresses and returns
a list object of a heatmap and the underlying plot data
}
\details{
\tabular{ll}{ Package: \tab ipv4heatmap\cr Type: \tab Package\cr
Version: \tab 1.0\cr Date: \tab 2014-07-29\cr License: \tab MIT \cr }
}
\author{
Bob Rudis

Maintainer: Bob Rudis
}
\keyword{package}

