\name{ipv4heatmap-package}
\alias{ipv4heatmap-package}
\alias{ipv4heatmap}
\docType{package}
\title{
What the package does (short line)
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab ipv4heatmap\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-29\cr
License: \tab What Licence is it under ?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
