% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ipv4heatmap}
\alias{ipv4heatmap}
\title{Return a ggplot2 object of a hilbert-space heatmap for a set of ip addys}
\usage{
ipv4heatmap(ips)
}
\arguments{
\item{ips}{character vector of ip addresses}
}
\value{
ggplot2 object which can be ggsave'd or manipulated further
}
\description{
Return a ggplot2 object of a hilbert-space heatmap for a set of ip addys
}

