% Generated by roxygen2 (4.0.1): do not edit by hand
\name{long2ip}
\alias{long2ip}
\title{Convert IP addresses in long integer format to dotted-decimal notation}
\usage{
long2ip(ip)
}
\arguments{
\item{ip}{input numeric (long int) vector}
}
\value{
vector of equivalent dotted-decimal IP addresses
}
\description{
Convert IP addresses in long integer format to dotted-decimal notation
}

